TTIS/requirements.txt
biopython>=1.78,<2.0
numpy>=1.20,<2.0